resolvePath() is now available directly to commands (#68)
I thought it would be nice to update the documentation and example to show that `resolvePath()` is available directly to commands now. Assuming you are ok with that update, I wasn't sure if you would want to retain the info about filesystemUtil being available to all commands as well, so I left it in a parenthetical note.