Removing failovermethod property
On rpm/yum/dnf-based distributions like RHEL, RL, Centos etc the `failovermethod` property is deprecated. `dnf` already actively doesn't support it anymore and throws a warning. Given that `yum` is on the way out as tooling and usage of `dnf` is recommended now, we might want to move to providing `dnf` instructions anyway.