Revise wording for Windows installation
The most important change is in the first sentence discussing Windows: the previous wording said to "Unzip the executable box.exe"...which could be misread, with someone thinking they need to unzip the exe itself (into its underlying files, which is possible), but that's not what's intended here at all.  

Then while I was here, I went ahead and tweaked the next couple of sentences to refine/clarify the points further, especially adding the SmartScreen prompt, since the current docs did propose how folks could run the exe by "double-clicking" on it, which is when Smartscreen kicks in (to be clear, it does NOT apply when instead running the exe from the command line, even for the first time after downloading it. That surprises me, too. I decided to clarify that here rather than in the text, in case anyone may feel the need to confirm that for themselves before accepting that part of this PR.)