Update apt-based GPG key process
A lot of `apt-key`'s functionality has been deprecated and it's considered bad practice to write into `trusted.gpg` or `trusted.gpg.d` as they key then essentially get added to every repository.

The above change ties the GPG key to the commandbox repo only.

It seems though that there's not a 100% established process of where to download and store the key at. People seem to be divided between `/usr/share/keyrings` and `/etc/apt/keyrings` . I went with the previous for all my setup.